MLModel/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.14.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.1.0
python-dotenv==1.0.0
werkzeug==3.0.0
requests==2.31.0
gunicorn==21.2.0
